Move nav link logic to useHeaderLinks hook
